requirements.txt
streamlit
openai
google-cloud-vision
gspread
python-dotenv
pandas